+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.uni</groupId>
     <artifactId>unit</artifactId>
-    <version>0.24.2</version>
+    <version>0.25</version>
     <dependencies>
         <!-- test framework -->
         <dependency>
